requirements.txt
pytest
allure-pytest
requests
faker
pika
mysql-connector-python
redis
